[backend]: seperate fileloader to a individual service; assume file loader behaves like database reader
