+++ Zed-dotnet/Zed-dotnet/Interfaces/IFileLoader.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using Zeddotnet.Models;
+
+namespace Zeddotnet.Interfaces
+{
+    /// <summary>
+    /// Interface of File Loader that declares functions
+    /// for load studies from json files
+    /// </summary>
+    public interface IFileLoader
+    {
+        Dictionary<string, Study> LoadStudies();
+    }
+}
