+++ Zed-dotnet/Zed-dotnet/Services/StudyManager.cs
@@ -11,50 +11,28 @@ namespace Zeddotnet.Services
 {
     /// <summary>
     /// The main class for managing Study
-    /// Contains all methods for performing basic loading and returning functions
+    /// Contains all methods for performing reading and returning studies 
     /// </summary>
     public class StudyManager : IStudyManager
     {
-        private Dictionary<string, Study> studies;
+        private readonly IFileLoader _fileLoader;
 
-        public StudyManager()
+        public StudyManager(IFileLoader fileLoader)
         {
-            if (studies == null || studies.Count == 0)
-            {
-                studies = LoadStudies();
-            }
-
-        }
-
-        // Read json files and deserialise to study object.
-        // Return a dictionary of studies
-        public Dictionary<string, Study> LoadStudies()
-        {
-
-            string root = "./Files";
-            string[] fileNames = Directory.GetFiles(root);
-
-            Dictionary<string, Study> tempStudies = new Dictionary<string, Study>();
-
-            foreach (string fileName in fileNames)
-            {
-                var jsonText = File.ReadAllText(fileName);
-                Study study = JsonConvert.DeserializeObject<Study>(jsonText);
-                tempStudies.Add(study.GUID, study);
-            }
-
-            return tempStudies;
+            _fileLoader = fileLoader;
         }
 
         // Return a dictionary of all studies stored in the File folder
         public Dictionary<string, Study> GetAll()
         {
-            return studies;
+            return _fileLoader.LoadStudies();
         }
 
         // Return a single study object found by guid.
         public Study GetStudyByGUID(string guid)
         {
+            var studies = _fileLoader.LoadStudies();
+
             if (studies.ContainsKey(guid) == true)
             {
                 return studies[guid];
