+++ Zed-dotnet/Zed-dotnet/Startup.cs
@@ -29,6 +29,7 @@ namespace Zed_dotnet
         {
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
             services.AddSingleton<IStudyManager, StudyManager>();
+            services.AddSingleton<IFileLoader, FileLoader>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
